moving fused_autodocs to interface.py, which remove the need to import fusedhelper.py and renders the dependencies on pandas and ipython optional. This fixes #67
